+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Cover `<UploadingJobsContextProvider>` component with tests (UIDATIMP-729)
 * Cover `<UploadingJobsDisplay>` component with tests (UIDATIMP-730)
 * Cover `FileExtensions` components with tests (UIDATIMP-973)
+* Cover `<JobProfiles` component with tests (UIDATIMP-974)
 
 ## [5.0.2](https://github.com/folio-org/ui-data-import/tree/v5.0.2) (2021-11-12)